Move panes to running state from debug panel
